chore: rename tauri drag region component
