+++ src/app/components/fullscreen/drag-region.tsx
